dom/file/test/test_lockedfile_lifetimes_nested.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<html>
<head>
  <title>File Handle Test</title>

  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

  <script type="text/javascript;version=1.7">
  function testSteps()
  {
    for each (let fileStorage in fileStorages) {
      let request = getFileHandle(fileStorage.key, "test.txt");
      request.onerror = errorHandler;
      request.onsuccess = grabEventAndContinueHandler;
      let event = yield;

      let fileHandle = event.target.result;
      fileHandle.onerror = errorHandler;

      let lockedFile = fileHandle.open();

      let lockedFile2;

      netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");
      let thread = Components.classes["@mozilla.org/thread-manager;1"]
                             .getService()
                             .currentThread;

      let eventHasRun;

      thread.dispatch(function() {
        eventHasRun = true;

        lockedFile2 = fileHandle.open();
      }, Components.interfaces.nsIThread.DISPATCH_NORMAL);

      while (!eventHasRun) {
        thread.processNextEvent(false);
      }

      ok(lockedFile2, "Non-null lockedFile2");

      continueToNextStep();
      yield;
    }

    finishTest();
    yield;
  }
  </script>
  <script type="text/javascript;version=1.7" src="helpers.js"></script>

</head>

<body onload="runTest();"></body>

</html>
